backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0  # MySQL驱动
gunicorn==20.1.0
requests==2.31.0
pandas==2.3.0
numpy==1.26.0
openai==0.27.8  # 如果使用AI查询功能
setuptools==68.0.0
wheel==0.41.2